requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#

click==6.6                # via pip-tools
first==2.0.1              # via pip-tools
Jinja2==2.8
MarkupSafe==0.23          # via jinja2
pip-tools==1.6.5
PyYAML==3.11
six==1.10.0               # via pip-tools
